backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
httpx==0.27.0
python-dotenv==1.0.1
sqlmodel==0.0.19
anthropic==0.39.0
apscheduler==3.10.4
